minor to fixes
