+++ e2e-tests/spring-rest-openapi-v2/src/test/java/org/evomaster/e2etests/spring/examples/wiremock/service/ExternalServiceTest.java
@@ -6,20 +6,39 @@ import com.foo.rest.examples.spring.wiremock.service.ServiceController;
 import com.github.tomakehurst.wiremock.WireMockServer;
 import com.github.tomakehurst.wiremock.core.WireMockConfiguration;
 import com.github.tomakehurst.wiremock.extension.responsetemplating.ResponseTemplateTransformer;
+import com.google.inject.Injector;
+import com.google.inject.Key;
+import com.google.inject.TypeLiteral;
 import org.evomaster.core.EMConfig;
+import org.evomaster.core.problem.external.service.ExternalServices;
 import org.evomaster.core.problem.rest.HttpVerb;
 import org.evomaster.core.problem.rest.RestIndividual;
+import org.evomaster.core.problem.rest.SampleType;
+import org.evomaster.core.problem.rest.resource.RestResourceCalls;
+import org.evomaster.core.problem.rest.service.ResourceManageService;
+import org.evomaster.core.problem.rest.service.ResourceRestMutator;
+import org.evomaster.core.problem.rest.service.ResourceSampler;
+import org.evomaster.core.problem.rest.service.RestResourceFitness;
+import org.evomaster.core.search.EvaluatedIndividual;
 import org.evomaster.core.search.Solution;
+import org.evomaster.core.search.service.Archive;
 import org.evomaster.e2etests.spring.examples.SpringTestBase;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.List;
+
 import static com.github.tomakehurst.wiremock.client.WireMock.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class ExternalServiceTest extends SpringTestBase {
     /**
-     * Ignore this test
+     * Ignore this test, incomplete
      */
 
     private static WireMockServer wireMockServer;
@@ -74,7 +93,7 @@ public class ExternalServiceTest extends SpringTestBase {
         DnsCacheManipulator.clearDnsCache();
     }
 
-    @Test
+    @Disabled
     public void testRunEM() throws Throwable {
         runTestHandlingFlakyAndCompilation(
                 "ExternalServiceEM",
@@ -86,5 +105,35 @@ public class ExternalServiceTest extends SpringTestBase {
 
                     assertHasAtLeastOne(solution, HttpVerb.GET, 200, "/api/wiremock/external", "true");
                 });
+    }
+
+    @Disabled
+    public void dummyTest() throws Throwable {
+
+        String[] args = new String[]{
+                "--createTests", "true",
+                "--seed", "42",
+                "--sutControllerPort", "" + controllerPort,
+                "--maxActionEvaluations", "1",
+                "--stoppingCriterion", "FITNESS_EVALUATIONS",
+                "--executiveSummary", "false",
+                "--expectationsActive" , "true",
+                "--outputFormat", "KOTLIN_JUNIT_5",
+                "--outputFolder", "target/em-tests/ExternalServiceEM"
+        };
+
+        Injector injector = init(Arrays.asList(args));
+
+        ExternalServices externalServices = injector.getInstance(ExternalServices.class);
+
+        // To check the url info at startup
+        RestResourceFitness ff = injector.getInstance(RestResourceFitness.class);
+        ResourceSampler resourceSampler = injector.getInstance(ResourceSampler.class); // check
+        RestIndividual restIndividual = resourceSampler.sample();
+
+        assertEquals(1, externalServices.getExternalServicesCount());
+        ff.calculateCoverage(restIndividual, Collections.emptySet());
+        assertEquals(2, externalServices.getExternalServicesCount());
+
     }
 }
